requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
httpx==0.27.0
beautifulsoup4==4.12.3
google-auth==2.30.0
google-auth-oauthlib==1.2.0
gspread==6.1.2
python-multipart==0.0.9
pydantic==2.7.4
pydantic-settings==2.3.4
python-dotenv==1.0.1
jsonschema==4.22.0
pandas==2.2.2
tldextract==5.1.2
aiofiles==23.2.1
loguru==0.7.2
